Create the V6 version for Biot's Problem in MATLAB
In this version, we apply the quadrature rule for the case of two tensors, to compute the integrals: (A(sigma_h + alpha*p_h*I),tau) and (A(sigma_h),wI).
We also compute directly the following integral: (A(alpha*p_h*I),wI)=(tr(A(alpha*p_h*I)),w).