Update MS_MF_MFE.m
We test if the program works well with the source term f being positive (as in the theory).
The errors are similar as in the case where f is negative (altough they seem to increase a little bit when refining the mesh).
But the graphics of the computed solution are much more in accordance with the graphics of the exact solution.