Create 'Test 1.C. (lambda=mu=1)--VERY BAD--
To obtain the initial data p0=p(x,0), we use the solution of the
pressure obtained by solving the discrete system for the stationary
eqs. of the flux problem when t=0:

A22=f'*(t\f);
q indep. term without delta_t
gDp=f'*(t\Pgp);

We obtain the rest of the initial data (compatible initial data)
by computing:

A) The solution for the displacement of the elasticity system when
t=0, considering implicitly the term alpha*grad p, thus:

sigma_e is considered instead of sigma.
f=-div(sigma_e) is considered.

B) The rest of the elasticity variables are computed as in the
elasticity problem (without A_sp^T), once u0 is obtained in the
previous step A).